refactor: isolate cloud webdriver scroll frame
